requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn

face_alignment
python_speech_features
numba
resampy
#pyaudio
soundfile==0.12.1
einops
configargparse

lpips==0.1.3
imageio-ffmpeg

transformers
edge_tts
flask
flask_sockets
opencv-python-headless
aiortc
aiohttp_cors

ffmpeg-python
omegaconf
diffusers
accelerate

librosa
openai
websockets==12.0

# Swagger文档相关依赖 (选择其中一个安装)
# aiohttp-swagger3  # 推荐：支持OpenAPI 3.0，更现代
aiohttp-swagger    # 备选：支持Swagger 2.0，更稳定
